+++ client/src/game/store.ts
@@ -87,6 +87,8 @@ class GameStore extends VuexModule implements GameState {
 
     selectionHelperID: string | null = null;
 
+    invertAlt = false;
+
     get selectedLayer(): string {
         return this.layers[this.selectedLayerIndex];
     }
@@ -488,6 +490,12 @@ class GameStore extends VuexModule implements GameState {
         if (data.sync) {
             socket.emit("Room.Info.Set.Locked", this.isLocked);
         }
+    }
+
+    @Mutation
+    setInvertAlt(data: { invertAlt: boolean; sync: boolean }): void {
+        this.invertAlt = data.invertAlt;
+        if (data.sync) socket.emit("Client.Options.Set", { invertAlt: data.invertAlt });
     }
 
     @Mutation